+++ Depth_of_Field/cuda/summed_area_table.cuh
@@ -1,4 +1,4 @@
 
 //coordinates must be swapped since not transposed again
-void summed_area_table(unsigned int* data, unsigned int N); 
+void summed_area_table(float* data, unsigned int N); 
 //Modified version of: https://github.com/thrust/thrust/blob/master/examples/summed_area_table.cu
